Made unread-count refresh more reliable.
Google Chrome will sometimes change the status icon for a page even when the
browser hasn't changed pages, when all that's changed is the part after the # in
the URL.  We now make sure to change the favicon even if the unread count for
the page hasn't changed.  This results in a favicon change every 30 seconds, but
this does not seem to have a performance penalty associated with it.  However,
it has not been thoroughly tested.